// project_nextjs_backup2/src/features/editor/types.ts
export const SUBMISSION_STAGES = ["submission", "review", "copyediting", "production"] as const;
export type SubmissionStage = (typeof SUBMISSION_STAGES)[number];

export const SUBMISSION_STATUSES = [
  "queued",
  "in_review",
  "in_copyediting",
  "in_production",
  "accepted",
  "scheduled",
  "published",
  "declined",
  "archived",
] as const;

export type SubmissionStatus = (typeof SUBMISSION_STATUSES)[number];
// ...
export const SUBMISSION_EDITOR_DECISION_EXTERNAL_REVIEW = 8; // Submission stage decision
export const SUBMISSION_EDITOR_DECISION_ACCEPT = 1; // Submission and review stages decision
export const SUBMISSION_EDITOR_DECISION_DECLINE = 4; // Submission and review stages decision
export const SUBMISSION_EDITOR_DECISION_INITIAL_DECLINE = 9; // PKP base decision
export const SUBMISSION_EDITOR_DECISION_PENDING_REVISIONS = 2; // Review stage decision
export const SUBMISSION_EDITOR_DECISION_RESUBMIT = 3; // Review stage decision
export const SUBMISSION_EDITOR_DECISION_SEND_TO_PRODUCTION = 7; // Editorial stage decision
export const SUBMISSION_EDITOR_DECISION_REVERT_DECLINE = 17; // PKP base decision
export const SUBMISSION_EDITOR_DECISION_NEW_ROUND = 16; // Review stage decision

// Editor Recommendation Constants
export const SUBMISSION_EDITOR_RECOMMEND_ACCEPT = 11;
export const SUBMISSION_EDITOR_RECOMMEND_PENDING_REVISIONS = 12;
export const SUBMISSION_EDITOR_RECOMMEND_RESUBMIT = 13;
export const SUBMISSION_EDITOR_RECOMMEND_DECLINE = 14;

export type EditorDecisionType =
  | typeof SUBMISSION_EDITOR_DECISION_EXTERNAL_REVIEW
  | typeof SUBMISSION_EDITOR_DECISION_ACCEPT
  | typeof SUBMISSION_EDITOR_DECISION_DECLINE
  | typeof SUBMISSION_EDITOR_DECISION_INITIAL_DECLINE
  | typeof SUBMISSION_EDITOR_DECISION_PENDING_REVISIONS
  | typeof SUBMISSION_EDITOR_DECISION_RESUBMIT
  | typeof SUBMISSION_EDITOR_DECISION_SEND_TO_PRODUCTION
  | typeof SUBMISSION_EDITOR_DECISION_REVERT_DECLINE
  | typeof SUBMISSION_EDITOR_DECISION_NEW_ROUND;

export type EditorRecommendationType =
  | typeof SUBMISSION_EDITOR_RECOMMEND_ACCEPT
  | typeof SUBMISSION_EDITOR_RECOMMEND_PENDING_REVISIONS
  | typeof SUBMISSION_EDITOR_RECOMMEND_RESUBMIT
  | typeof SUBMISSION_EDITOR_RECOMMEND_DECLINE;